+++ backend/tests/test_service_exports.py
@@ -75,7 +75,7 @@ class Tests(util.TestCase):
 
     @mock.patch('mora.service.exports.os.path.isdir', lambda x: True)
     @mock.patch('mora.service.exports.os.path.isfile', lambda x: True)
-    @mock.patch('mora.service.exports.flask.send_file')
+    @mock.patch('mora.service.exports.flask.send_from_directory')
     def test_get_export_file_returns_file(self, mock_send_file):
         """Ensure we return a file if found"""