+++ NEWS.rst
@@ -4,7 +4,10 @@ In development
 New features
 ------------
 
-
+* #39244: Added root filter to orgunit search
+* #39244: Unified the paged_filtered_get and paged_get methods, thus changing
+  the order of paged lists throughout MO. The order is now by UUID instead of
+  by user_key / bvn.
 
 Version 1.10.1, 2020-10-26
 ==========================